Convert telemetry service to typescript and simplify it
