--- frontend/src/app/TelemetryService.js
@@ -1,50 +1,0 @@
-import { ApplicationInsights } from "@microsoft/applicationinsights-web";
-import { ReactPlugin } from "@microsoft/applicationinsights-react-js";
-
-let reactPlugin = null;
-let appInsights = null;
-
-/**
- * Create the App Insights Telemetry Service
- * @return {{reactPlugin: ReactPlugin, appInsights: Object, initialize: Function}} - Object
- */
-const createTelemetryService = () => {
-  /**
-   * Initialize the Application Insights class
-   * @param {string} instrumentationKey - Application Insights Instrumentation Key
-   * @param {Object} browserHistory - client's browser history, supplied by the withRouter HOC
-   * @return {void}
-   */
-  const initialize = (instrumentationKey, browserHistory) => {
-    if (!browserHistory) {
-      throw new Error("Could not initialize Telemetry Service");
-    }
-    if (!instrumentationKey) {
-      throw new Error("Instrumentation key not provided in");
-    }
-
-    reactPlugin = new ReactPlugin();
-
-    appInsights = new ApplicationInsights({
-      config: {
-        instrumentationKey: process.env.REACT_APP_APPINSIGHTS_KEY,
-        extensions: [reactPlugin],
-        loggingLevelConsole: 2,
-        disableFetchTracking: false,
-        enableAutoRouteTracking: true,
-        loggingLevelTelemetry: 2,
-        maxBatchInterval: 0,
-        extensionConfig: {
-          [reactPlugin.identifier]: { history: browserHistory },
-        },
-      },
-    });
-
-    appInsights.loadAppInsights();
-  };
-
-  return { reactPlugin, appInsights, initialize };
-};
-
-export const ai = createTelemetryService();
-export const getAppInsights = () => appInsights;
